cutout_dev/requirements.txt
amqp<2.0,>=1.4.9
anyjson==0.3.3
asn1crypto==0.22.0
billiard
celery<4.0,>=3.1.15
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
Django==1.9.5
django-celery>=3.1
django-debug-toolbar==1.9.1
enum34==1.1.6
h5py==2.7.0
idna==2.6
ipaddress==1.0.18
kombu<3.1,>=3.0.37
numpy==1.13.3
psycopg2==2.7.3.1
pycparser==2.18
pyodbc==4.0.24
pytz==2017.2
six==1.10.0
sqlparse==0.2.4
vtk==8.1.1
